% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{dt_to_xts}
\alias{dt_to_xts}
\title{Convert data.table to xts time series}
\usage{
dt_to_xts(dt)
}
\arguments{
\item{dt}{A data.table with date and OHLC columns}
}
\value{
An xts object with OHLC data
}
\description{
Convert data.table to xts time series
}
