% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{getMode}
\alias{getMode}
\title{Get the mode (most frequent value)}
\usage{
getMode(v)
}
\arguments{
\item{v}{A vector}
}
\value{
The most frequent value in the vector
}
\description{
Get the mode (most frequent value)
}
