% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{getTimeFrameMins}
\alias{getTimeFrameMins}
\title{Calculate timeframe in minutes}
\usage{
getTimeFrameMins(dat)
}
\arguments{
\item{dat}{A data.table with date column}
}
\value{
Numeric value representing minutes between bars
}
\description{
Calculate timeframe in minutes
}
