% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{mergeHighLowTimes}
\alias{mergeHighLowTimes}
\title{Merge data from different timeframes}
\usage{
mergeHighLowTimes(
  ask,
  lgTP = "h1",
  smTP = "m1",
  cols = "Close",
  fun,
  ...,
  onlyComplete = TRUE
)
}
\arguments{
\item{ask}{Price data}

\item{lgTP}{Larger timeframe}

\item{smTP}{Smaller timeframe}

\item{cols}{Columns to process}

\item{fun}{Function to apply}

\item{...}{Additional arguments}

\item{onlyComplete}{Only use complete candles}
}
\description{
Merge data from different timeframes
}
