% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotReturnsHistogram}
\alias{plotReturnsHistogram}
\title{Plot returns histogram}
\usage{
plotReturnsHistogram(bt)
}
\arguments{
\item{bt}{Backtest object returned from backtest() function}
}
\value{
A plotly histogram
}
\description{
Creates a histogram of trade returns
}
