% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotReturnsTimeline}
\alias{plotReturnsTimeline}
\title{Plot returns over time}
\usage{
plotReturnsTimeline(bt)
}
\arguments{
\item{bt}{Backtest object returned from backtest() function}
}
\value{
A plotly bar chart
}
\description{
Creates a bar chart of returns over time
}
