% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotWinLossPie}
\alias{plotWinLossPie}
\title{Plot win/loss pie chart}
\usage{
plotWinLossPie(bt)
}
\arguments{
\item{bt}{Backtest object returned from backtest() function}
}
\value{
A plotly pie chart
}
\description{
Creates a pie chart showing the distribution of winning and losing trades
}
