% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{xts_to_dt}
\alias{xts_to_dt}
\title{Convert xts time series to data.table}
\usage{
xts_to_dt(ts)
}
\arguments{
\item{ts}{An xts time series object}
}
\value{
A data.table with date column and original data
}
\description{
Convert xts time series to data.table
}
