% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeTesterR-package.R
\docType{package}
\name{TradeTesterR-package}
\alias{TradeTesterR}
\alias{TradeTesterR-package}
\title{TradeTesterR}
\description{
Core package documentation and central imports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alb3rtazzo/TradeTesterR}
  \item Report bugs at \url{https://github.com/alb3rtazzo/TradeTesterR/issues}
}

}
\author{
\strong{Maintainer}: Alberto Pallotta \email{pallottaalberto@gmail.com}

}
\keyword{internal}
